requirements.txt
streamlit
numpy
pillow
scikit-learn==1.6.1  # IMPORTANT: Use your exact version
tensorflow
requests